Don't create compendia for organisms that don't have a QN target.
