+++ workers/data_refinery_workers/processors/test_qn_reference.py
@@ -18,6 +18,7 @@ from data_refinery_common.models import (
     ProcessorJobDatasetAssociation
 )
 from data_refinery_common.models.organism import Organism
+from data_refinery_common.utils import get_most_recent_qn_target_for_organism
 from data_refinery_workers.processors import qn_reference, smasher, utils
 
 
@@ -73,7 +74,7 @@ class QNRefTestCase(TransactionTestCase):
             exsa.sample = sample
             exsa.save()
 
-        
+
         dataset = Dataset()
         dataset.data = {"12345": ["1", "2", "3", "4", "5", "6"]}
         dataset.aggregate_by = "ALL"
@@ -91,7 +92,7 @@ class QNRefTestCase(TransactionTestCase):
         self.assertTrue(os.path.exists(final_context['target_file']))
         self.assertEqual(os.path.getsize(final_context['target_file']), 559)
 
-        target = utils.get_most_recent_qn_target_for_organism(homo_sapiens)
+        target = get_most_recent_qn_target_for_organism(homo_sapiens)
         self.assertEqual(target.sha1, '7e97c077dbbda3eb5fbbf0e306a151eb6d233135')
 
         ###
@@ -121,7 +122,7 @@ class QNRefTestCase(TransactionTestCase):
         self.assertEqual(final_context['merged_qn']['1'][0], -0.4379488528812934)
         self.assertEqual(final_context['original_merged']['1'][0], -0.576210936113982)
 
-        ## 
+        ##
         # Test via management command
         ##
 
@@ -139,4 +140,4 @@ class QNRefTestCase(TransactionTestCase):
         self.assertTrue('Target file' in stdout)
         path = stdout.split('\n')[0].split(':')[1].strip()
         self.assertTrue(os.path.exists(path))
-        self.assertEqual(path, utils.get_most_recent_qn_target_for_organism(homo_sapiens).absolute_file_path)
+        self.assertEqual(path, get_most_recent_qn_target_for_organism(homo_sapiens).absolute_file_path)
